third_party/WebKit/LayoutTests/fast/forms/select/popup-menu-update-from-element.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script src="../resources/common.js"></script>
<script src="../resources/picker-common.js"></script>
</head>
<body>
<select id="menu">
  <option selected>foo</option>
  <option>bar</option>
  <option hidden>baz</option>
  <optgroup label="qux">
      <option>garply</option>
  </optgroup>
</select>
<script>
var menuElement = document.getElementById('menu');
var picker;

function openPickerErrorCallback() {
    testFailed('picker didn\'t open')
    finishJSTest();
}
openPicker(menu, test1, openPickerErrorCallback);

function test1() {
    picker = window.internals.pagePopupWindow.global.picker;

    function onUpdate() {
        setTimeout(function() {
            shouldBeEqualToString('picker._selectElement.options[0].style.color', 'rgb(123, 0, 0)');
            picker.removeListener("didUpdate", onUpdate);
            test2();
        });
    }
    picker.on("didUpdate", onUpdate);
    menuElement.style.color = 'rgb(123, 0, 0)';
}

function test2() {
    menuElement.innerHTML = '<option>FOO</option><optgroup label="BAR"><option selected>BAZ</option></optgroup>';

    picker.on("didUpdate", function () {
        setTimeout(function () {
            shouldBe('picker._selectElement.options.length', '2');
            shouldBeEqualToString('picker._selectElement.children[0].label', 'FOO');
            shouldBeEqualToString('picker._selectElement.children[1].tagName', 'OPTGROUP');
            shouldBeEqualToString('picker._selectElement.children[1].label', 'BAR');
            shouldBeEqualToString('picker._selectElement.children[1].children[0].label', 'BAZ');
            finishJSTest();
        }, 0);
    });
}
</script>
</body>
</html>
